// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const ReviewStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type ReviewStatus = (typeof ReviewStatus)[keyof typeof ReviewStatus]


export const PricingType = {
  FREE: 'FREE',
  PREMIUM: 'PREMIUM'
} as const

export type PricingType = (typeof PricingType)[keyof typeof PricingType]


export const PurchaseType = {
  BUY: 'BUY',
  RENT: 'RENT'
} as const

export type PurchaseType = (typeof PurchaseType)[keyof typeof PurchaseType]


export const PurchaseStatus = {
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED'
} as const

export type PurchaseStatus = (typeof PurchaseStatus)[keyof typeof PurchaseStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const SubscriptionStatus = {
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
  CANCELLED: 'CANCELLED'
} as const

export type SubscriptionStatus = (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus]


export const SubscriptionPlan = {
  MONTHLY: 'MONTHLY',
  YEARLY: 'YEARLY'
} as const

export type SubscriptionPlan = (typeof SubscriptionPlan)[keyof typeof SubscriptionPlan]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BANNED: 'BANNED',
  SUSPENDED: 'SUSPENDED',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const ContentType = {
  MOVIE: 'MOVIE',
  SERIES: 'SERIES'
} as const

export type ContentType = (typeof ContentType)[keyof typeof ContentType]